+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 **1.8.8-RC8**
+* Print out `API book compiled` when running in dev mode #668
 * It shall ignore Exception raised during app shutdown process #667
 * It shall clear OSGL Config's internal cache on app restart #666
 * update osgl library dependencies